fixed bugs introduced by merge

+++ src/Table.cpp
@@ -66,7 +66,7 @@ bool Table::Contains(std::string element) {
     return std::find(table_[h].begin(), table_[h].end(), getFingerprint(h)) != table_[h].end();
 }
 
-bool Table::Contains(int element) {
+bool Table::Contains(uint64_t element) {
     uint64_t h = getHash(element);
     return std::find(table_[h].begin(), table_[h].end(), getFingerprint(h)) != table_[h].end();
 }
@@ -82,8 +82,8 @@ bool Table::Remove(std::string element) {
 }
 
 bool Table::Remove(uint64_t element) {
-    size_t h = getHash(element);
-    std::list<int>::iterator it = std::find(table_[h].begin(), table_[h].end(), getFingerprint(h));
+    uint64_t h = getHash(element);
+    std::list<uint16_t>::iterator it = std::find(table_[h].begin(), table_[h].end(), getFingerprint(h));
     if (it != table_[h].end()) {
         table_[h].erase(it);
         return true;
